LayoutTests/fast/js/dfg-cfg-simplify-phantom-get-local-on-same-block-set-local-expected.txt

Tests that attempts by the DFG simplification to short-circuit a Phantom to a GetLocal on a variable that is SetLocal'd in the same block, and where the predecessor block(s) make no mention of that variable, do not result in crashes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS thingy(o) is 42
PASS successfullyParsed is true

TEST COMPLETE
